+++ src/main/java/com/upc/repository/ServicioXAdviserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.upc.model.ServicioXAdviser;
 
 @Repository
-public interface ServicioXAdviserRepository extends CrudRepository<ServicioXAdviser, Long> {
+public interface ServicioXAdviserRepository extends CrudRepository<ServicioXAdviser, Integer> {
 
 }
