+++ src/main/java/com/upc/repository/StudentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.upc.model.Student;
 
 @Repository
-public interface StudentRepository extends CrudRepository<Student, Long> {
+public interface StudentRepository extends CrudRepository<Student, Integer> {
 
 }
